fw-4255, move dictionary factories to factories dir
