+++ firstvoices/backend/tests/factories/__init__.py
@@ -1,1 +1,2 @@
 from .base import *  # noqa F401
+from .dictionary import *  # noqa F401
